Enhance timetable models and repositories with ID management
- Added an ID field to DMTimeSlot and DMWeekDay classes, enabling automatic ID generation.
- Updated StudentTimetableRepository and WeekDayRepository to use Short as the ID type instead of Long, improving consistency across the application.
- Adjusted the method signature in StudentTimetableRepository to reflect the new ID type for dayId.

+++ src/main/kotlin/com/attendanceio/api/repository/timetable/StudentTimetableRepository.kt
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository
 interface StudentTimetableRepository : JpaRepository<DMStudentTimetable, Long> {
     fun findByStudentId(studentId: Long): List<DMStudentTimetable>
     fun findByStudentIdAndSemesterId(studentId: Long, semesterId: Long): List<DMStudentTimetable>
-    fun findByStudentIdAndSemesterIdAndDayId(studentId: Long, semesterId: Long, dayId: Long): List<DMStudentTimetable>
+    fun findByStudentIdAndSemesterIdAndDayId(studentId: Long, semesterId: Long, dayId: Short): List<DMStudentTimetable>
     
     @Modifying
     @Query("DELETE FROM DMStudentTimetable st WHERE st.student.id = :studentId")
